feat: release package as a library (#80)
* feat: release package as a library

* fix: export ApiClient

* chore: update README
