+++ index.ts
@@ -1,4 +1,5 @@
 export {
+    ApiClient,
     BugSplatApiClient,
     VersionsApiClient,
     SymbolsApiClient,
@@ -6,3 +7,5 @@ export {
     UploadableFile,
     GZippedSymbolFile
 } from '@bugsplat/js-api-client';
+
+export { uploadSymbolFiles } from './src/upload';
